Move MonoBitmapManager and factory classes into monobimap-manager module
