+++ libs/monobitmap-manager/src/main/kotlin/mono/bitmap/manager/MonoBitmapManager.kt
@@ -1,8 +1,9 @@
-package mono.graphics.bitmap
+package mono.bitmap.manager
 
-import mono.graphics.bitmap.drawable.LineDrawable
-import mono.graphics.bitmap.drawable.RectangleDrawable
-import mono.graphics.bitmap.drawable.TextDrawable
+import mono.bitmap.manager.factory.LineBitmapFactory
+import mono.bitmap.manager.factory.RectangleBitmapFactory
+import mono.bitmap.manager.factory.TextBitmapFactory
+import mono.graphics.bitmap.MonoBitmap
 import mono.shape.shape.AbstractShape
 import mono.shape.shape.Group
 import mono.shape.shape.Line
@@ -13,8 +14,6 @@ import mono.shape.shape.Text
 /**
  * A model class which manages and caches bitmap of shapes.
  * Cache-hit when both id and version of the shape valid in the cache, otherwise, cache-miss.
- *
- * Call [invalidate] to clean up cache of a shape.
  */
 class MonoBitmapManager {
     private val idToBitmapMap: MutableMap<String, VersionizedBitmap> = mutableMapOf()
@@ -26,13 +25,13 @@ class MonoBitmapManager {
         }
 
         val bitmap = when (shape) {
-            is Rectangle -> RectangleDrawable.toBitmap(shape.bound.size, shape.extra)
-            is Text -> TextDrawable.toBitmap(
+            is Rectangle -> RectangleBitmapFactory.toBitmap(shape.bound.size, shape.extra)
+            is Text -> TextBitmapFactory.toBitmap(
                 shape.bound.size,
                 shape.renderableText.getRenderableText(),
                 shape.extra
             )
-            is Line -> LineDrawable.toBitmap(
+            is Line -> LineBitmapFactory.toBitmap(
                 shape.reducedJoinPoints,
                 shape.extra,
             )
@@ -42,12 +41,6 @@ class MonoBitmapManager {
         } ?: return null
         idToBitmapMap[shape.id] = VersionizedBitmap(shape.version, bitmap)
         return bitmap
-    }
-
-    fun hasShape(shape: AbstractShape): Boolean = shape.id in idToBitmapMap
-
-    fun invalidate(shape: AbstractShape) {
-        idToBitmapMap.remove(shape.id)
     }
 
     private fun getCacheBitmap(shape: AbstractShape): MonoBitmap? =
