+++ libs/monobitmap-manager/src/test/kotlin/mono/bitmap/manager/factory/RectangleBitmapFactoryTest.kt
@@ -1,4 +1,4 @@
-package mono.graphics.bitmap.drawable
+package mono.bitmap.manager.factory
 
 import mono.graphics.geo.Rect
 import mono.shape.shape.Rectangle
@@ -6,13 +6,13 @@ import kotlin.test.Test
 import kotlin.test.assertEquals
 
 /**
- * A test for [RectangleDrawable]
+ * A test for [RectangleBitmapFactory]
  */
-class RectangleDrawableTest {
+class RectangleBitmapFactoryTest {
     @Test
     fun testToBitmap() {
         val shape = Rectangle(Rect.byLTWH(10, 10, 5, 5))
-        val bitmap = RectangleDrawable.toBitmap(shape.bound.size, shape.extra)
+        val bitmap = RectangleBitmapFactory.toBitmap(shape.bound.size, shape.extra)
         assertEquals(shape.bound.width, bitmap.size.width)
         assertEquals(shape.bound.height, bitmap.size.height)
         assertEquals(
