+++ libs/monobitmap-manager/src/test/kotlin/mono/bitmap/manager/factory/TextBitmapFactoryTest.kt
@@ -1,4 +1,4 @@
-package mono.graphics.bitmap.drawable
+package mono.bitmap.manager.factory
 
 import mono.graphics.geo.Rect
 import mono.shape.shape.Text
@@ -6,14 +6,14 @@ import kotlin.test.Test
 import kotlin.test.assertEquals
 
 /**
- * A test for [TextDrawableTest]
+ * A test for [TextBitmapFactory]
  */
-class TextDrawableTest {
+class TextBitmapFactoryTest {
     @Test
     fun testToBitmap() {
         val text = Text(Rect.byLTWH(0, 0, 7, 5))
         text.setText("012345678\nabc")
-        val bitmap = TextDrawable.toBitmap(
+        val bitmap = TextBitmapFactory.toBitmap(
             text.bound.size,
             text.renderableText.getRenderableText(),
             text.extra
